@@ -18,7 +18,7 @@ buildscript {
 
 
 ext {
-    compileSdkVersion = 26
+    compileSdkVersion = 28
     minSdkVersion = 16
     targetSdkVersion = 22
     versionCode = 28
