@@ -1,16 +1,12 @@
 package com.xiaoniu.cleanking.ui.main.activity;
 
-import android.app.ActivityManager;
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.KeyEvent;
-import android.widget.Button;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
@@ -34,22 +30,14 @@ import com.xiaoniu.cleanking.ui.main.widget.BottomBar;
 import com.xiaoniu.cleanking.ui.main.widget.BottomBarTab;
 import com.xiaoniu.cleanking.ui.main.widget.SPUtil;
 import com.xiaoniu.cleanking.utils.AndroidUtil;
+import com.xiaoniu.cleanking.utils.DbHelper;
 import com.xiaoniu.cleanking.utils.StatisticsUtils;
-import com.xiaoniu.cleanking.utils.encypt.Base64;
 import com.xiaoniu.cleanking.utils.prefs.NoClearSPHelper;
 import com.xiaoniu.cleanking.utils.update.UpdateAgent;
 
-import java.io.UnsupportedEncodingException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.crypto.BadPaddingException;
-import javax.crypto.Cipher;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.NoSuchPaddingException;
-import javax.crypto.spec.SecretKeySpec;
 import javax.inject.Inject;
 
 import butterknife.BindView;
@@ -166,30 +154,7 @@ public class MainActivity extends BaseActivity<MainPresenter> {
 
          btn_wjgl.setOnClickListener(v-> startActivity(FileManagerHomeActivity.class));
 
-
-        try {
-            String s2 = "DdAt8o/mcQ08hIcdQ0p/0hp9IHf95AaQ2xkDDmqVR8U=";
-            String s = "225e8C70688fD76Ec5C01A392302320A".toUpperCase();
-            SecretKeySpec secretKeySpec = new SecretKeySpec(s.getBytes("utf-8"), "AES");
-            Cipher instance = Cipher.getInstance("AES/ECB/PKCS5Padding");
-            instance.init(2, secretKeySpec);
-            final byte[] decode = Base64.decode(s2, 0);
-
-            String s1 = new String(instance.doFinal(decode), "utf-8");
-            System.out.println(s1);
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        } catch (NoSuchPaddingException e) {
-            e.printStackTrace();
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
-        } catch (InvalidKeyException e) {
-            e.printStackTrace();
-        } catch (BadPaddingException e) {
-            e.printStackTrace();
-        } catch (IllegalBlockSizeException e) {
-            e.printStackTrace();
-        }
+        DbHelper.copyDb();
     }
 
 
@@ -203,52 +168,6 @@ public class MainActivity extends BaseActivity<MainPresenter> {
         }else if(id==R.id.btn_clean_video){
             startActivity(new Intent(this,CleanVideoManageActivity.class));
         }
-    }
-    private void cleanMemory() {
-        ActivityManager am = (ActivityManager) getSystemService(Context.ACTIVITY_SERVICE);
-        List<ActivityManager.RunningAppProcessInfo> infoList = am.getRunningAppProcesses();
-        List<ActivityManager.RunningServiceInfo> serviceInfos = am.getRunningServices(100);
-
-        long beforeMem = getAvailMemory(MainActivity.this);
-        Log.d(TAG, "-----------before memory info : " + beforeMem);
-        int count = 0;
-        if (infoList != null) {
-            for (int i = 0; i < infoList.size(); ++i) {
-                ActivityManager.RunningAppProcessInfo appProcessInfo = infoList.get(i);
-                Log.d(TAG, "process name : " + appProcessInfo.processName);
-                //importance 该进程的重要程度  分为几个级别，数值越低就越重要。
-                Log.d(TAG, "importance : " + appProcessInfo.importance);
-
-                // 一般数值大于RunningAppProcessInfo.IMPORTANCE_SERVICE的进程都长时间没用或者空进程了
-                // 一般数值大于RunningAppProcessInfo.IMPORTANCE_VISIBLE的进程都是非可见进程，也就是在后台运行着
-                if (appProcessInfo.importance > ActivityManager.RunningAppProcessInfo.IMPORTANCE_VISIBLE) {
-                    String[] pkgList = appProcessInfo.pkgList;
-                    for (int j = 0; j < pkgList.length; ++j) {//pkgList 得到该进程下运行的包名
-                        Log.d(TAG, "It will be killed, package name : " + pkgList[j]);
-                        am.killBackgroundProcesses(pkgList[j]);
-                        count++;
-                    }
-                }
-
-            }
-        }
-
-        long afterMem = getAvailMemory(MainActivity.this);
-        Log.d(TAG, "----------- after memory info : " + afterMem);
-        Toast.makeText(MainActivity.this, "clear " + count + " process, "
-                + (afterMem - beforeMem) + "M", Toast.LENGTH_LONG).show();
-    }
-
-    //获取可用内存大小
-    private long getAvailMemory(Context context) {
-        // 获取android当前可用内存大小
-        ActivityManager am = (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
-        ActivityManager.MemoryInfo mi = new ActivityManager.MemoryInfo();
-        am.getMemoryInfo(mi);
-        //mi.availMem; 当前系统的可用内存
-        //return Formatter.formatFileSize(context, mi.availMem);// 将获取的内存大小规格化
-        Log.d(TAG, "可用内存---->>>" + mi.availMem / (1024 * 1024));
-        return mi.availMem / (1024 * 1024);
     }
 
     /**
@@ -290,7 +209,6 @@ public class MainActivity extends BaseActivity<MainPresenter> {
     @Override
     public void inject(ActivityComponent activityComponent) {
         activityComponent.inject(this);
-        mPresenter.saveCacheFiles();
     }
 
     private void initFragments() {
